pr(script-generator): move CaseEnumSerializer to the library module
